+++ src/components/forecasts/MenuCard.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from "react";
+import React from "react";
 import {
   Container,
   // Row,
@@ -9,12 +9,11 @@ import {
   // ListGroupItem,
   Accordion,
   // ToggleButtonGroup
-  Modal,
 } from "react-bootstrap";
 import "./forecast.css";
 import "leaflet/dist/leaflet.css";
-import electricity2015 from "./biddingzones/2015electricity";
-import { Line } from "react-chartjs-2";
+
+import ChartFor2015Electricity from "./Chart2015Electricity";
 
 /*
  * TODO MAKE MAP ZOOM AND PAN
@@ -26,95 +25,11 @@ const menuCard = (setFeature, feature, zone, lat, lng) => {
     const TextFeature = () => {
       // ZONE FEATURE START
       if (feature === "zones") {
-        const Electricity2015button = (props) => {
-          const [show, setShow] = useState(false);
-          const [chartData, setChartData] = useState({});
-          const handleClose = () => setShow(false);
-          const handleShow = () => setShow(true);
-          const chart = () => {
-            let data = [];
-            if (zone && countries2015electricity.includes(zone)) {
-              data = electricity2015[zone].map((integer) => Number(integer));
-            }
-            setChartData({
-              labels: data,
-              datasets: [
-                {
-                  label: zone,
-                  data,
-                  backgroundColor: ["rgba(75, 192, 192, 0.6"],
-                  borderWidth: 4,
-                  pointRadius: 0,
-                },
-              ],
-            });
-          };
-          useEffect(() => {
-            chart();
-          }, []);
-          if (countries2015electricity.includes(zone)) {
-            return (
-              <>
-                <br />
-                <Button variant="info" onClick={handleShow}>
-                  2015 Electricity for {zone}
-                </Button>
-                <Modal centered show={show} onHide={handleClose} size="lg">
-                  <Modal.Header closeButton>
-                    <Modal.Title>
-                      Electricity prize per hour for 2015 for {zone}
-                    </Modal.Title>
-                  </Modal.Header>
-                  <Modal.Body>
-                    <Line
-                      data={chartData}
-                      options={{
-                        responsive: true,
-                        scales: {
-                          yAxes: [
-                            {
-                              ticks: {
-                                autoSkip: true,
-                                maxTicksLimit: 0,
-                              },
-                              gridLines: {
-                                display: false,
-                              },
-                            },
-                          ],
-                          xAxes: [
-                            {
-                              gridLines: { display: false },
-                            },
-                          ],
-                        },
-                      }}
-                    />
-                  </Modal.Body>
-                  <Modal.Footer>
-                    <Button variant="secondary" onClick={handleClose}>
-                      Close
-                    </Button>
-                  </Modal.Footer>
-                </Modal>
-              </>
-            );
-          } else if (zone) {
-            return (
-              <>
-                <br /> Zone Selected: {zone}
-              </>
-            );
-          } else {
-            return null;
-          }
-        };
-
         return (
           <>
             <div className="text-center">
               <h5>Zones</h5>Select a Zone to see more information.
-              <Electricity2015button />
+              <ChartFor2015Electricity zone={zone} />
               <br />
             </div>
           </>
